Surface::set_mime_data() - take ownership of a Vec to avoid a 'static lifetime requirement
Taking an AsRef([u8]) with a 'static requirement means that callers
cannot read image buffers (say, from a file) into the heap and pass
them on to set_mime_data().

Instead, now we take ownership of the data buffer.  This is the same
semantics of cairo_surface_set_mime_data(), which takes ownership of
the buffer.

Fixes https://github.com/gtk-rs/cairo/issues/228