Return a Result instead of a plain Context from Context::new and the from_ pointer functions
Similar to what we do for Surfaces, we now check the status of a
cairo_t coming from the C world.

Context::clone is special in that it doesn't check the status of the
cloned context.  It is just meant to increase the refcount, whether
the context is in an error state or not.